% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_i.R
\name{html_dependency_phosphor}
\alias{html_dependency_phosphor}
\title{HTML dependency for Phosphor}
\usage{
html_dependency_phosphor()
}
\value{
An \code{htmltools::htmlDependency()} object.
}
\description{
HTML dependency for Phosphor
}
