% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_icon}
\alias{search_icon}
\title{Search Icons}
\usage{
search_icon(pattern, view_results = interactive())
}
\arguments{
\item{pattern}{Pattern to search for in icon's name.}

\item{view_results}{View icons found in viewer.}
}
\value{
a character vector invisibly.
}
\description{
Search Icons
}
\examples{
search_icon("calendar")
search_icon("text")
}
