% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_i.R
\name{html_dependency_phosphor}
\alias{html_dependency_phosphor}
\title{HTML dependency for Phosphor Icons}
\usage{
html_dependency_phosphor()
}
\value{
An \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} object.
}
\description{
Allow to explicitly load dependency for using Phosphor icons.
}
\note{
Dependency is automatically loaded when using \code{\link[=ph_i]{ph_i()}}, but in some case,
like when using \code{icon} argument in some function, you might need to call \code{html_dependency_phosphor()} to make icons appears.
}
\examples{

library(shiny)
library(phosphoricons)

ui <- navbarPage(
  title = "Phosphor Icons",
  header = list(
    html_dependency_phosphor()
  ),
  tabPanel("Home", icon = ph_i("house")),
  tabPanel("Parameters"),
  tabPanel("Results")
)

server <- function(input, output, session) {
  
}

if (interactive())
  shinyApp(ui, server)
}
