% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{ph}
\alias{ph}
\title{Phosphor Icons}
\usage{
ph(
  name,
  type = c("light", "regular", "thin", "bold", "fill"),
  fill = "#333",
  height = "1.33em",
  width = NULL,
  ...
)
}
\arguments{
\item{name}{Name of the icon to use.}

\item{type}{Type of icon.}

\item{fill}{Fill color for the icon.}

\item{height, width}{Height and width in valid CSS unit.}

\item{...}{Attributes passed to the SVG tag.}
}
\value{
An SVG tag.
}
\description{
Create a \verb{<svg></svg>} tag to render an icon.
}
