% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_icon}
\alias{search_icon}
\title{Search Icons}
\usage{
search_icon(pattern, output = c("viewer", "console"))
}
\arguments{
\item{pattern}{Pattern to search for in icon's name.}

\item{output}{View results in the viewer or in the console.}
}
\value{
If \code{output = "console"} a character vector, if \code{output = "viewer"} a \code{shiny.tag}.
}
\description{
Search Icons
}
\examples{
search_icon("calendar")
search_icon("text")
}
